// store/creatorManager/creatorManager.types.ts
export enum CreatorStages {
  NOT_STARTED = 'NOT_STARTED',
  CHOOSE_TEMPLATE = 'CHOOSE_TEMPLATE',
  SET_ROLES = 'SET_ROLES',
  SET_VAULTS = 'SET_VAULTS',
  SHARE_DISTRIBUTION = 'SHARE_DISTRIBUTION',
  SET_PROFIT_SWITCH = 'SET_PROFIT_SWITCH',
  PROVIDE_EMAILS = 'PROVIDE_EMAILS',
  PROVIDE_DETAILS = 'PROVIDE_DETAILS',
  CONFIRM = 'CONFIRM',
  CONFIRM_MULTIPLE = 'CONFIRM_MULTIPLE',
  FINISEHD = 'FINISEHD',
  SET_NAME = 'SET_NAME',
  CONFIRM_SELF_MANAGED = 'CONFIRM_SELF_MANAGED',
  SELF_MANAGED_KYC = 'SELF_MANAGED_KYC'
}

export type CreatorManager = {
  currentStage: CreatorStages
  stagesList: CreatorStages[]
}